Add gazebo_proxy.cpp
