+++ ros_gz_sim/src/gz_simulation_interfaces/gazebo_proxy.hpp
@@ -19,15 +19,14 @@
 #include <gz/msgs/serialized_map.pb.h>
 #include <gz/msgs/stringmsg_v.pb.h>
 
+#include <memory>
+#include <string>
+
 #include <gz/math/Pose3.hh>
 #include <gz/sim/EntityComponentManager.hh>
 #include <gz/sim/Util.hh>
 #include <gz/transport/Node.hh>
-#include <iostream>
-#include <memory>
-#include <optional>
 #include <rclcpp/rclcpp.hpp>
-#include <stdexcept>
 
 namespace ros_gz_sim
 {
@@ -36,95 +35,20 @@ namespace gz_simulation_interfaces
 class GazeboProxy
 {
 public:
-  struct State
-  {
-    gz::math::Pose3d pose;
-    gz::math::Vector3d linear_velocity;
-    gz::math::Vector3d angular_velocity;
-  };
+  GazeboProxy(const std::string world_name, std::shared_ptr<rclcpp::Node> ros_node);
 
-  GazeboProxy(const std::string world_name, std::shared_ptr<rclcpp::Node> ros_node)
-  : world_name_(world_name), ros_node_(ros_node), gz_node_(std::make_shared<gz::transport::Node>())
-  {
-    if (!this->InitializeGazeboConnection()) {
-      throw std::runtime_error("Could not initialize Gazebo connection");
-    }
+  bool InitializeGazeboConnection();
 
-    // TODO(azeey) Consider adding the UserCommands system if not already present.
-    // TODO(azeey) Wait for critical services to be available (e.g. /world/*/create,
-    // /world/*/control) Request the initial state of the world. This will block until Gazebo is
-    // initialized
-    gz::msgs::SerializedStepMap reply;
-    bool result;
-    if (!this->gz_node_->Request(this->PrefixTopic("state"), 30000, reply, result)) {
-      RCLCPP_ERROR(
-        ros_node->get_logger(),
-        "Simulation interface timed out while waiting for Gazebo to initialize");
-      return;
-    } else {
-      if (!result) {
-        RCLCPP_ERROR(
-          ros_node->get_logger(),
-          "Simulation interface encountered an error while synchronizing state with Gazebo");
-        return;
-      } else {
-        this->UpdateStateFromMsg(reply);
+  std::string PrefixTopic(const char * topic) const;
+  void UpdateStateFromMsg(const gz::msgs::SerializedStepMap & msg);
 
-        std::cout << "Subscribe to " << this->PrefixTopic("state") << "\n";
-        // Listen to the "state" topic to get periodic updates.
-        if (!this->gz_node_->Subscribe(
-              this->PrefixTopic("state"), &GazeboProxy::UpdateStateFromMsg, this)) {
-          RCLCPP_ERROR(ros_node->get_logger(), "Subscribing to continues state updates failed");
-        }
-      }
-    }
-  }
+  uint64_t Iterations() const;
 
-  bool InitializeGazeboConnection()
-  {
-    gz::msgs::StringMsg_V worlds_msg;
-    bool result;
-    if (this->gz_node_->Request(
-          "gazebo/worlds", GazeboProxy::kGzServiceTimeout, worlds_msg, result)) {
-      if (result && !worlds_msg.data().empty()) {
-        this->world_name_ = worlds_msg.data(0);
-        return true;
-      }
-    }
-    return false;
-  }
-  std::string PrefixTopic(const char * topic) { return "world/" + this->world_name_ + "/" + topic; }
-  void UpdateStateFromMsg(const gz::msgs::SerializedStepMap & msg)
-  {
-    std::lock_guard<std::mutex> lk(this->stateSyncMutex_);
-    this->ecm_.SetState(msg.state());
-    this->ecm_.ClearRemovedComponents();
-    this->ecm_.ClearNewlyCreatedEntities();
-    this->ecm_.ProcessRemoveEntityRequests();
-    this->world_stats_ = msg.stats();
+  bool Paused() const;
 
-    // TODO(azeey) Consider using a condition variable to notify services that there is new data so
-    // as to avoid using stale data
-  }
+  void WithEcm(std::function<void(gz::sim::EntityComponentManager &)> f);
 
-  uint64_t Iterations()
-  {
-    std::lock_guard<std::mutex> lk(this->stateSyncMutex_);
-    return this->world_stats_.iterations();
-  }
-  bool Paused()
-  {
-    std::lock_guard<std::mutex> lk(this->stateSyncMutex_);
-    return this->world_stats_.paused();
-  }
-
-  void WithEcm(std::function<void(gz::sim::EntityComponentManager &)> f)
-  {
-    std::lock_guard<std::mutex> lk(this->stateSyncMutex_);
-    f(this->ecm_);
-  }
-
-  std::shared_ptr<gz::transport::Node> GzNode() { return this->gz_node_; }
+  std::shared_ptr<gz::transport::Node> GzNode();
 
   static constexpr unsigned int kGzServiceTimeout{5000};
