+++ ros_gz_sim/src/gz_simulation_interfaces/utils.cpp
@@ -24,23 +24,6 @@ namespace ros_gz_sim
 {
 namespace gz_simulation_interfaces
 {
-bool ConvertState(
-  const GazeboProxy::State & gz_state, simulation_interfaces::msg::EntityState & state)
-{
-  // TODO(azeey) Fill in header
-  state.pose.position.x = gz_state.pose.X();
-  state.pose.position.y = gz_state.pose.Y();
-  state.pose.position.z = gz_state.pose.Z();
-
-  state.pose.orientation.x = gz_state.pose.Rot().X();
-  state.pose.orientation.y = gz_state.pose.Rot().Y();
-  state.pose.orientation.z = gz_state.pose.Rot().Z();
-  state.pose.orientation.w = gz_state.pose.Rot().W();
-
-  // TODO(azeey) Add support for twists and accelerations
-  // TODO(azeey) Implement error checking and setting error message
-  return true;
-}
 
 void ConvertPose(const gz::math::Pose3d & gz_pose, geometry_msgs::msg::Pose & ros_pose)
 {
@@ -61,7 +44,8 @@ geometry_msgs::msg::Pose ConvertPose(const gz::math::Pose3d & gz_pose)
   return ros_pose;
 }
 
-void ConvertPose(const geometry_msgs::msg::Pose & ros_pose, gz::math::Pose3d & gz_pose){
+void ConvertPose(const geometry_msgs::msg::Pose & ros_pose, gz::math::Pose3d & gz_pose)
+{
   gz_pose.Pos().X() = ros_pose.position.x;
   gz_pose.Pos().Y() = ros_pose.position.y;
   gz_pose.Pos().Z() = ros_pose.position.z;
@@ -72,20 +56,22 @@ void ConvertPose(const geometry_msgs::msg::Pose & ros_pose, gz::math::Pose3d & g
   gz_pose.Rot().W() = ros_pose.orientation.w;
 }
 
-gz::math::Pose3d  ConvertPose(const geometry_msgs::msg::Pose & ros_pose) {
+gz::math::Pose3d ConvertPose(const geometry_msgs::msg::Pose & ros_pose)
+{
   gz::math::Pose3d gz_pose;
   ConvertPose(ros_pose, gz_pose);
   return gz_pose;
 }
 
-
-void ConvertVector3(const geometry_msgs::msg::Vector3 & ros_v, gz::math::Vector3d & gz_v) {
+void ConvertVector3(const geometry_msgs::msg::Vector3 & ros_v, gz::math::Vector3d & gz_v)
+{
   gz_v.X() = ros_v.x;
   gz_v.Y() = ros_v.y;
   gz_v.Z() = ros_v.z;
 }
 
-gz::math::Vector3d ConvertVector3(const geometry_msgs::msg::Vector3 & ros_v) {
+gz::math::Vector3d ConvertVector3(const geometry_msgs::msg::Vector3 & ros_v)
+{
   gz::math::Vector3d gz_v;
   ConvertVector3(ros_v, gz_v);
   return gz_v;
