+++ ros_gz_sim/src/gz_simulation_interfaces/utils.hpp
@@ -18,7 +18,6 @@
 #include <gz/math/Pose3.hh>
 #include <simulation_interfaces/msg/entity_state.hpp>
 
-#include "gazebo_proxy.hpp"
 #include "geometry_msgs/msg/pose.hpp"
 #include "geometry_msgs/msg/vector3.hpp"
 
@@ -26,14 +25,11 @@ namespace ros_gz_sim
 {
 namespace gz_simulation_interfaces
 {
-bool ConvertState(
-  const GazeboProxy::State & gz_proxy, simulation_interfaces::msg::EntityState & state);
-
 void ConvertPose(const gz::math::Pose3d & gz_pose, geometry_msgs::msg::Pose & ros_pose);
 geometry_msgs::msg::Pose ConvertPose(const gz::math::Pose3d & gz_pose);
 
 void ConvertPose(const geometry_msgs::msg::Pose & ros_pose, gz::math::Pose3d & gz_pose);
-gz::math::Pose3d  ConvertPose(const geometry_msgs::msg::Pose & ros_pose);
+gz::math::Pose3d ConvertPose(const geometry_msgs::msg::Pose & ros_pose);
 
 void ConvertVector3(const geometry_msgs::msg::Vector3 & ros_v, gz::math::Vector3d & gz_v);
 gz::math::Vector3d ConvertVector3(const geometry_msgs::msg::Vector3 & ros_v);
